preset-0.60 + tweaks
