+++ packages/haul-core/jest.config.js
@@ -1,5 +1,10 @@
+const path = require('path');
+
 module.exports = {
   displayName: '@haul-bundler/core',
   testPathIgnorePatterns: ['__fixtures__/.*'],
   testRegex: '/__tests__/.*\\.(test|spec)\\.(j|t)sx?$',
+  moduleNameMapper: {
+    '^jest/(.*)': path.join(__dirname, '../../jest/$1'),
+  },
 };
