+++ packages/haul-core/package.json
@@ -22,8 +22,10 @@
     "@hapi/hapi": "^18.3.1",
     "@hapi/inert": "^5.2.0",
     "@hapi/joi": "^15.0.3",
+    "@haul-bundler/basic-bundle-webpack-plugin": "^0.6.0",
     "@haul-bundler/core-legacy": "^0.6.0",
     "@haul-bundler/inspector-events": "^0.6.0",
+    "@haul-bundler/ram-bundle-webpack-plugin": "^0.6.0",
     "ansi-fragments": "^0.2.1",
     "dedent": "^0.7.0",
     "escape-string-regexp": "^2.0.0",
@@ -31,6 +33,7 @@
     "image-size": "^0.7.4",
     "ink": "^2.2.0",
     "loader-utils": "^1.2.3",
+    "lodash.get": "^4.4.2",
     "lodash.throttle": "^4.1.1",
     "node-fetch": "^2.6.0",
     "open": "^6.0.0",
@@ -53,6 +56,7 @@
     "@types/hapi__joi": "^15.0.1",
     "@types/image-size": "^0.7.0",
     "@types/loader-utils": "^1.1.3",
+    "@types/lodash.get": "^4.4.6",
     "@types/lodash.throttle": "^4.1.6",
     "@types/node-fetch": "^2.3.6",
     "@types/react": "^16.8.20",
