+++ packages/haul-preset-0.60/README.md
@@ -1,0 +1,26 @@
+# @haul-bundler/preset-0.60
+
+[![Version][version]][package]   
+
+[![PRs Welcome][prs-welcome-badge]][prs-welcome]
+[![MIT License][license-badge]][license]
+[![Chat][chat-badge]][chat]
+[![Code of Conduct][coc-badge]][coc]
+
+Project preset with support for __React Native 0.60.x__ for Haul - Webpack-based React Native bundler.
+
+You can read moe about Haul here: https://github.com/callstack/haul.
+
+<!-- badges (common) -->
+
+[license-badge]: https://img.shields.io/npm/l/@haul-bundler/preset-0.60.svg?style=flat-square
+[license]: https://opensource.org/licenses/MIT
+[prs-welcome-badge]: https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square
+[prs-welcome]: http://makeapullrequest.com
+[coc-badge]: https://img.shields.io/badge/code%20of-conduct-ff69b4.svg?style=flat-square
+[coc]: https://github.com/callstack/haul/blob/master/CODE_OF_CONDUCT.md
+[chat-badge]: https://img.shields.io/badge/chat-discord-brightgreen.svg?style=flat-square&colorB=7289DA&logo=discord
+[chat]: https://discord.gg/zwR2Cdh
+
+[version]: https://img.shields.io/npm/v/@haul-bundler/preset-0.60.svg?style=flat-square
+[package]: https://www.npmjs.com/package/@haul-bundler/preset-0.60
