+++ packages/haul-preset-0.60/src/defaultConfig.ts
@@ -110,16 +110,6 @@ export default function getDefaultConfig(
       }),
     ],
     resolve: {
-      alias:
-        process.env.NODE_ENV === 'production'
-          ? {}
-          : {
-              /**
-               * Latest `react-proxy` version does not contain try/catches from
-               * commit 981815dca250373619138c9f5aadf12295cf1b3f.
-               */
-              'react-proxy': '@zamotany/react-proxy',
-            },
       plugins: [
         /**
          * React Native uses a module system called Haste
