+++ packages/haul-ram-bundle-webpack-plugin/package.json
@@ -12,13 +12,11 @@
   "license": "MIT",
   "bugs": "https://github.com/callstack/haul/issues",
   "peerDependencies": {
+    "@haul-bundler/core": "^0.6.0",
     "metro": "^0.53.1",
     "terser": "^3.17.0",
     "webpack": "^4.30.0",
     "webpack-sources": "^1.3.0"
-  },
-  "dependencies": {
-    "@haul-bundler/core": "^0.6.0"
   },
   "devDependencies": {
     "@babel/core": "^7.4.4",
